Ticket closing shows auto-contact list
On the ticket close form, we now show the list of people who will automatically be contacted, via email.  We also show a list of the people for whom we do not have email addresses.  This second list of people will need to be contacted seperately.  We want the user to record log entries for each of the people they contact.

Updates #231