Render sort url as a string
I have a current $url object that we are constantly
modifying as we iterate over all the fields.  This is fine,
except objects get passed by reference.  So after the full
loop, all of the assignments we did to the sort link helper
now have exactly the same $url object.  We must render the
url out as a string each time before doing more
modifications.

Updates #260