Use flat space instead of SRID 4326
For the purposes of the ticket map clusters, the distance calculation is
just not important or sensitve enough to work about the difference
between flat space and a curved earth.

The curved projection distance calculations are very, very, very slow.
It would take about a week to reindex the geoclusters.  Using flat
space, the reindex completes in under an hour.

IN the old version of mysql all the spatial calculations were done in
flat space, anyway.  So this keeps doing what we were already doing in
the previous version.

Updates #380