Fixed initial setup errors when no dept exists
Person, Department, and Category are all interrelated.  Errors were preventing creating the initial
person, department, and category...as they each depended on the others already existing.

This fixes the expectation that a department exists or that it has people.  This allows the very
first department and category to be created.

Updates #384