New involvement time query for reporting
This new query tries to sum up the time each person spends with a ticket by joining the ticketHistory table on itself.  The start time is when they were assigned the ticket, and the end time is when they either closed it or assigned it to someone else.

This should be a drop in replacement for the existing reporting system.

Updates #396