Search: ignore results not in the database
It is possible that the search index gets out of sync with the database.  As bots scan over all pages of all search results, tickets/unknown errors are triggered.

This can be resolved by doing a full index on the search engine more often.  However, these types of errors do not need to trigger a notification in Graylog.