Open311: fixed field name for phone numbers
The fieldname for a person's phone number ended up being different when searching for a person by phone number versus saving their phone number.

This commit makes it so the search and saving of a person's phone number uses the same field name.  This adopts the fieldname of 'phone'.

Updates #406